requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
